+++ src/store/Store.ts
@@ -1,51 +1,16 @@
-import { Identifier } from "@babel/types";
 import path = require("path");
-import {
-  languages,
-  TextEditor,
-  Uri,
-  window,
-  TextDocument,
-  Range,
-  workspace,
-} from "vscode";
-import {
-  CssModuleExtensions,
-  CSS_MODULE_EXTENSIONS,
-  MODULE_EXTENSIONS,
-} from "../constants";
-import { ParserResult } from "../parser/v2/ts";
+import { languages, TextEditor, Uri, window, Range, workspace } from "vscode";
+import { CssModuleExtensions, CSS_MODULE_EXTENSIONS } from "../constants";
 import * as fsg from "fast-glob";
-import { CssParserResult, parseCss, Selector } from "../parser/v2/css";
 import { promises as fs_promises } from "node:fs";
 import Settings from "../settings";
 import { Parser } from "../parser/Parser";
 import { DiagnosticsProvider } from "../providers/ts/diagnostics";
 import { normalizePath } from "../path-utils";
 
-type FileName = string;
-type StyleIdentifier = Identifier["name"];
-
-export type StyleReferences = {
-  style_references: Map<
-    StyleIdentifier,
-    {
-      uri: string;
-    }
-  >;
-};
-
-interface CSSReferences {
-  references: Set<FileName>;
-}
-
 // Full file path of the active opened file
-type CssModules = Map<string, CssParserResult & CSSReferences>;
-type Experimental_CssModules = Map<string, string>;
-export type ParsedResult = Map<FileName, ParserResult & StyleReferences>;
-export type TsModules = ParsedResult;
+type CssModules = Map<string, string>;
 
-export type TsConfigMap = Map<string, TsConfig>;
 export type TsConfig = {
   compilerOptions: {
     baseUrl?: string;
@@ -56,14 +21,15 @@ export type TsConfig = {
   baseDir: string;
 };
 
+export type TsConfigMap = Map<string, TsConfig>;
+
 export type IgnoreDiagnostis = Map<
   string, // Selector
   Range // Range of the Selector
 >;
+
 export class Store {
-  protected _cssModules: CssModules = new Map();
-  public experimental_cssModules: Experimental_CssModules = new Map();
-  protected _tsModules: TsModules = new Map();
+  public cssModules: CssModules = new Map();
   /** Root path of the workspace */
   protected _workSpaceRoot: string | undefined;
   static diagonisticCollection =
@@ -98,41 +64,6 @@ export class Store {
     this._workSpaceRoot = v;
   }
 
-  /**
-   * Map of all the css/scss files in the workspace
-   */
-  public get cssModules(): CssModules {
-    return this._cssModules;
-  }
-
-  /**
-   * Set the source css files on activation
-   * TODO: store only modules
-   */
-  private async setCssModules() {
-    const uri = window.activeTextEditor?.document?.uri;
-    if (uri) {
-      if (!this.workSpaceRoot) {
-        const _uri = workspace.getWorkspaceFolder(uri)?.uri;
-        const workspaceRoot = _uri?.fsPath;
-        this.workSpaceRoot = workspaceRoot;
-      }
-      const glob = `**/*.{${CSS_MODULE_EXTENSIONS.map((e) =>
-        e.replace(".", "")
-      ).join(",")}}`;
-      const files = await fsg(glob, {
-        cwd: this.workSpaceRoot,
-        ignore: ["node_modules", "build", "dist", "coverage"],
-        absolute: true,
-      });
-      await Promise.all(
-        files.map(async (v) => {
-          await this.storeCssParserResult(v);
-        })
-      );
-    }
-  }
-
   private async experimental_setCssModules() {
     const uri = window.activeTextEditor?.document?.uri;
     if (uri) {
@@ -149,47 +80,9 @@ export class Store {
         ignore: ["node_modules", "build", "dist", "coverage"],
         absolute: true,
       });
-      files.forEach((file) => this.experimental_cssModules.set(file, file));
+      files.forEach((file) => this.cssModules.set(file, file));
     }
   }
-
-  /**
-   * Map of all the ts/tsx files in the workspace
-   */
-  public get tsModules(): TsModules {
-    return this._tsModules;
-  }
-
-  // private async storeTsParserResult({
-  //   filePath: file,
-  //   content,
-  //   parserFactory,
-  // }: {
-  //   filePath: string;
-  //   parserFactory: Parser;
-  //   content: string;
-  // }) {
-  //   const result = await parserFactory.parse({
-  //     filePath: normalizePath(file),
-  //     content,
-  //   });
-  //   if (result) {
-  //     this.tsModules.set(file, {
-  //       ...result.parsedResult,
-  //       style_references: result.style_references,
-  //     });
-  //     for (const [, value] of result.style_references) {
-  //       if (this.cssModules.has(value.uri)) {
-  //         const module = this.cssModules.get(value.uri);
-  //         if (module) {
-  //           module.references.add(file);
-  //           this.cssModules.set(value.uri, module);
-  //         }
-  //       }
-  //     }
-  //   }
-  //   return result;
-  // }
 
   private async saveTsConfigAutomatically() {
     try {
@@ -242,22 +135,9 @@ export class Store {
           path.extname(f.fsPath) as CssModuleExtensions
         )
       ) {
-        await this.storeCssParserResult(f.path);
+        this.cssModules.set(f.path, f.path);
       }
     });
-  }
-
-  public async storeCssParserResult(module: string) {
-    try {
-      const result = await parseCss(module);
-      const cached = this._cssModules.get(module);
-      if (result) {
-        this._cssModules.set(normalizePath(module), {
-          ...result,
-          references: cached?.references ?? new Set(),
-        });
-      }
-    } catch (e) {}
   }
 
   resolveCssModuleAlias(source: string): string | undefined {
@@ -330,57 +210,12 @@ export class Store {
     return this.activeTextEditor.document;
   }
 
-  public async bootStrap() {
-    try {
-      if (this.activeTextEditor.document.isDirty) {
-        return;
-      }
-      // await this.saveTsConfig();
-      if (!this.cssModules.size) {
-        await this.setCssModules();
-      }
-      await this.saveTsConfigAutomatically();
-      if (!this.tsModules.size) {
-        // Don't do this for every tsx file but rather only for active document until code lens is figured out.
-        // await this.setTsModules();
-      }
-      const document = this.activeTextEditor.document;
-      const filePath = document.uri.fsPath;
-      const uri = document.uri;
-
-      if (
-        CSS_MODULE_EXTENSIONS.includes(
-          path.extname(filePath) as CssModuleExtensions
-        )
-      ) {
-        this.storeCssParserResult(filePath);
-      } else {
-        // const content = document.getText();
-        const workspaceRoot = workspace.getWorkspaceFolder(uri)?.uri.fsPath;
-        // const parserFactory = new Parser({
-        //   workspaceRoot,
-        //   tsConfig: this.tsConfig,
-        //   baseDir: Settings.baseDir,
-        //   cssModules: this.cssModules,
-        // });
-        // await this.storeTsParserResult({
-        //   filePath,
-        //   content,
-        //   parserFactory,
-        // });
-        if (Settings.diagnostics) {
-          return this.provideDiagnostics();
-        }
-      }
-    } catch (e) {}
-  }
-
   public async experimental_BootStrap() {
     try {
       if (this.activeTextEditor.document.isDirty) {
         return;
       }
-      if (!this.experimental_cssModules.size) {
+      if (!this.cssModules.size) {
         await this.experimental_setCssModules();
       }
       await this.saveTsConfigAutomatically();
@@ -399,46 +234,26 @@ export class Store {
   }
 
   /**
-   * Get all selectors for a given Identifier
-   * @param identfier Identifier['name']
-   * @returns Map<string,Selector>
-   */
-  public getStyleReferenceByIdentifier(identfier: Identifier["name"]) {
-    return this.getActiveTsModule()?.style_references.get(identfier);
-  }
-
-  public getActiveTsModule() {
-    const filePath = this.activeTextEditor.document.uri.fsPath;
-    return this.tsModules.get(filePath);
-  }
-
-  public getTsModuleByPath(file: string) {
-    return this.tsModules.get(file);
-  }
-
-  /**
    * Flushes the entire storage on deactivation or opening a new workspace
    */
   public flushStorage() {
     this.workSpaceRoot = undefined;
-    this._cssModules = new Map();
-    this._tsModules = new Map();
     this.tsConfig = new Map();
+    this.cssModules = new Map();
   }
 
-  private provideDiagnostics() {
-    const parsedResult = this.getActiveTsModule();
+  private async provideDiagnostics() {
     const activeFileDir = path.parse(
       this.activeTextEditor.document.uri.fsPath
     ).dir;
     const activeFileUri = this.activeTextEditor.document.uri;
-    if (Settings.diagnostics) {
+    if (Settings.diagnostics && this.parser) {
       this.diagnosticsProvider = new DiagnosticsProvider({
-        parsedResult,
+        parser: this.parser,
         activeFileDir,
         activeFileUri,
       });
-      this.diagnosticsProvider.runDiagnostics();
+      await this.diagnosticsProvider.runDiagnostics();
       this.diagnosticsProvider.provideDiagnostics();
       return this.diagnosticsProvider.getDiagnostics();
     }
@@ -453,4 +268,5 @@ export class Store {
     this.provideDiagnostics();
   }
 }
+
 export default new Store();