+++ src/test/suite/extension.test.ts
@@ -31,6 +31,7 @@ import {
   ReferenceCodeLens,
   ReferenceCodeLensProvider,
 } from "../../providers/css/codelens";
+import { parseCss } from "../../parser/v2/css";
 const examplesLocation = "../../../examples/";
 
 function setWorskpaceFolder(app: string) {
@@ -102,7 +103,7 @@ suite("Extension Test Suite", async () => {
       test("should provide definitions when definition command is triggered at a relavent position [Class identifier]", async () => {
         const document = await workspace.openTextDocument(TestComponentUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
 
         const definition = new DefnitionProvider();
         const position = new Position(6, 34);
@@ -115,7 +116,7 @@ suite("Extension Test Suite", async () => {
       test("should not provide definitions if the command is triggered at a irrelavent position [no class identifier]", async () => {
         const document = await workspace.openTextDocument(TestComponentUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
 
         const definition = new DefnitionProvider();
         const position = new Position(4, 34);
@@ -129,7 +130,7 @@ suite("Extension Test Suite", async () => {
       test("should go to the correct definition content when definition is triggered on suffixed/nested selectors", async () => {
         const document = await workspace.openTextDocument(TestComponentUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
 
         const definition = new DefnitionProvider();
         const suffixResult = (await definition.provideDefinition(
@@ -157,7 +158,7 @@ suite("Extension Test Suite", async () => {
       test("should create a hovering content on hover at relavent position [Class selctor idenftier]", async () => {
         const document = await workspace.openTextDocument(TestComponentUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
 
         const hover = new HoverProvider();
         const position = new Position(6, 34);
@@ -174,7 +175,7 @@ suite("Extension Test Suite", async () => {
       test("should not create a hovering content on hover at irrelavent position [Class selctor idenftier]", async () => {
         const document = await workspace.openTextDocument(TestComponentUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
 
         const hover = new HoverProvider();
         const position = new Position(4, 34);
@@ -186,7 +187,7 @@ suite("Extension Test Suite", async () => {
       test("should show the correct hover content when hover on suffix selectors", async () => {
         const document = await workspace.openTextDocument(TestComponentUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
 
         const definition = new HoverProvider();
         const result = (await definition.provideHover(
@@ -200,7 +201,7 @@ suite("Extension Test Suite", async () => {
       test("should work for camel case selector values", async () => {
         const document = await workspace.openTextDocument(TestComponentUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
 
         const definition = new HoverProvider();
         const result = (await definition.provideHover(
@@ -216,7 +217,7 @@ suite("Extension Test Suite", async () => {
       test("should provide correct number of completions when triggered at the relavent position", async () => {
         const document = await workspace.openTextDocument(TestComponentUri);
 
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
 
         const completion = new SelectorsCompletionProvider();
         const position = new Position(6, 31);
@@ -247,7 +248,7 @@ suite("Extension Test Suite", async () => {
         writeFileSync(cssDocument.uri.fsPath, enc.encode(contents));
 
         const document = await workspace.openTextDocument(TestComponentUri);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
 
         const completion = new SelectorsCompletionProvider();
         const position = new Position(6, 31);
@@ -282,7 +283,7 @@ suite("Extension Test Suite", async () => {
         const enc = new TextEncoder();
         writeFileSync(cssDocument.uri.fsPath, enc.encode(contents));
 
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const completion = new SelectorsCompletionProvider();
         writeFileSync(cssDocument.uri.fsPath, enc.encode(contents));
         const position = new Position(6, 31);
@@ -307,7 +308,7 @@ suite("Extension Test Suite", async () => {
       test("should provide import completions on accessing styles identifier", async () => {
         const document = await workspace.openTextDocument(AutoImportComponent);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const completion = new ImportCompletionProvider();
         const position = new Position(6, 31);
         const list = await completion.provideCompletionItems(
@@ -321,7 +322,7 @@ suite("Extension Test Suite", async () => {
       test("should not provide import completions of already imported module on accessing styles identifier", async () => {
         const document = await workspace.openTextDocument(AutoImportComponent1);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const completion = new ImportCompletionProvider();
         const position = new Position(6, 31);
         const list = await completion.provideCompletionItems(
@@ -337,14 +338,14 @@ suite("Extension Test Suite", async () => {
       test("should provide diagnostics for missing selector", async () => {
         const document = await workspace.openTextDocument(TestComponentUri);
         await window.showTextDocument(document);
-        const diagnostics = await StorageInstance.bootStrap();
+        const diagnostics = await StorageInstance.experimental_BootStrap();
         assert.equal(diagnostics?.length, 1);
       });
 
       test("should provide diagnostics for in correct css module import", async () => {
         const document = await workspace.openTextDocument(DiagnosticComponent);
         await window.showTextDocument(document);
-        const diagnostics = await StorageInstance.bootStrap();
+        const diagnostics = await StorageInstance.experimental_BootStrap();
         assert.equal(diagnostics?.length, 2);
       });
 
@@ -358,7 +359,7 @@ suite("Extension Test Suite", async () => {
         );
         const document = await workspace.openTextDocument(DynamicClasses);
         await window.showTextDocument(document);
-        const diagnostics = await StorageInstance.bootStrap();
+        const diagnostics = await StorageInstance.experimental_BootStrap();
         assert.equal(diagnostics?.length, 0);
       });
     });
@@ -373,10 +374,11 @@ suite("Extension Test Suite", async () => {
     test("should include normal selectors [no relationship or bound to any rules]", async () => {
       const document = await workspace.openTextDocument(SelectorCssModule);
       await window.showTextDocument(document);
-      await StorageInstance.bootStrap();
-      const node = StorageInstance.cssModules.get(
+      await StorageInstance.experimental_BootStrap();
+      const source_css_file = StorageInstance.cssModules.get(
         normalizePath(SelectorCssModule)
       );
+      const node = await parseCss(source_css_file ?? "");
       assert.notEqual(node, undefined);
       const selectors = node!.selectors;
       assert.equal(
@@ -391,10 +393,11 @@ suite("Extension Test Suite", async () => {
     test("should include  selectors from mixins and media queries", async () => {
       const document = await workspace.openTextDocument(SelectorCssModule);
       await window.showTextDocument(document);
-      await StorageInstance.bootStrap();
-      const node = StorageInstance.cssModules.get(
+      await StorageInstance.experimental_BootStrap();
+      const source_css_file = StorageInstance.cssModules.get(
         normalizePath(SelectorCssModule)
       );
+      const node = await parseCss(source_css_file ?? "");
       assert.notEqual(node, undefined);
       const selectors = node!.selectors;
       assert.equal(selectors.get("flex-row")?.selector, "flex-row");
@@ -406,10 +409,11 @@ suite("Extension Test Suite", async () => {
     test("should include selectors from placeholders", async () => {
       const document = await workspace.openTextDocument(SelectorCssModule);
       await window.showTextDocument(document);
-      await StorageInstance.bootStrap();
-      const node = StorageInstance.cssModules.get(
+      await StorageInstance.experimental_BootStrap();
+      const source_css_file = StorageInstance.cssModules.get(
         normalizePath(SelectorCssModule)
       );
+      const node = await parseCss(source_css_file ?? "");
       assert.notEqual(node, undefined);
       const selectors = node!.selectors;
       assert.equal(selectors.get("place-holder")?.selector, "place-holder");
@@ -418,10 +422,11 @@ suite("Extension Test Suite", async () => {
     test("should include suffixed selectors at any depth", async () => {
       const document = await workspace.openTextDocument(SelectorCssModule);
       await window.showTextDocument(document);
-      await StorageInstance.bootStrap();
-      const node = StorageInstance.cssModules.get(
+      await StorageInstance.experimental_BootStrap();
+      const source_css_file = StorageInstance.cssModules.get(
         normalizePath(SelectorCssModule)
       );
+      const node = await parseCss(source_css_file ?? "");
       assert.notEqual(node, undefined);
       const selectors = node!.selectors;
       assert.equal(
@@ -441,10 +446,11 @@ suite("Extension Test Suite", async () => {
     test("should include camelCased suffixed selectors", async () => {
       const document = await workspace.openTextDocument(SelectorCssModule);
       await window.showTextDocument(document);
-      await StorageInstance.bootStrap();
-      const node = StorageInstance.cssModules.get(
+      await StorageInstance.experimental_BootStrap();
+      const source_css_file = StorageInstance.cssModules.get(
         normalizePath(SelectorCssModule)
       );
+      const node = await parseCss(source_css_file ?? "");
       assert.notEqual(node, undefined);
       const selectors = node!.selectors;
       assert.equal(selectors.get("camelCase")?.selector, "camelCase");
@@ -474,19 +480,25 @@ suite("Extension Test Suite", async () => {
       test("provide completions for css variables across files", async () => {
         const document = await workspace.openTextDocument(AppCssUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new CssVariablesCompletion();
         const position = new Position(6, 31);
-        const result = provider.provideCompletionItems(document, position);
+        const result = await provider.provideCompletionItems(
+          document,
+          position
+        );
         assert.equal((result?.items.length ?? 0) > 1, true);
       });
       test("completion items should resolve item to `var(${name})` if no `var` key word exists", async () => {
         const document = await workspace.openTextDocument(AppCssUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new CssVariablesCompletion();
         const position = new Position(6, 31);
-        const result = provider.provideCompletionItems(document, position);
+        const result = await provider.provideCompletionItems(
+          document,
+          position
+        );
         assert.equal(
           result?.items[0].insertText?.toString().includes("var"),
           true
@@ -496,10 +508,13 @@ suite("Extension Test Suite", async () => {
       test("completions items should not resolve to `var${name}` when var keyword exists", async () => {
         const document = await workspace.openTextDocument(AppCssUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new CssVariablesCompletion();
         const position = new Position(46, 14);
-        const result = provider.provideCompletionItems(document, position);
+        const result = await provider.provideCompletionItems(
+          document,
+          position
+        );
         assert.equal(
           result?.items[0].insertText?.toString().includes("var"),
           false
@@ -508,10 +523,13 @@ suite("Extension Test Suite", async () => {
       test("dont provide completions for css variables from same file", async () => {
         const document = await workspace.openTextDocument(IndexCssUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new CssVariablesCompletion();
         const position = new Position(6, 31);
-        const result = provider.provideCompletionItems(document, position);
+        const result = await provider.provideCompletionItems(
+          document,
+          position
+        );
         assert.equal(result?.items.length, 0);
       });
     });
@@ -519,20 +537,20 @@ suite("Extension Test Suite", async () => {
       test("provide definitions for variables across different files", async () => {
         const document = await workspace.openTextDocument(AppCssUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new CssDefinitionProvider();
         const position = new Position(40, 21);
-        const result = provider.provideDefinition(document, position);
+        const result = await provider.provideDefinition(document, position);
         assert.equal(result.length > 0, true);
       });
 
       test("dont provide definitions for variables within the same file", async () => {
         const document = await workspace.openTextDocument(IndexCssUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new CssDefinitionProvider();
         const position = new Position(23, 20);
-        const result = provider.provideDefinition(document, position);
+        const result = await provider.provideDefinition(document, position);
         assert.equal(result.length === 0, true);
       });
     });
@@ -540,9 +558,9 @@ suite("Extension Test Suite", async () => {
       test("provide color information for variables across different files", async () => {
         const document = await workspace.openTextDocument(AppCssUri);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new CssDocumentColorProvider();
-        const result = provider.provideDocumentColors(document);
+        const result = await provider.provideDocumentColors(document);
         assert.equal(result.length > 0, true);
       });
     });
@@ -551,7 +569,7 @@ suite("Extension Test Suite", async () => {
       test("provide references for a selector at a given position", async () => {
         const document = await workspace.openTextDocument(TestCssModulePath);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new ReferenceProvider();
         const result = await provider.provideReferences(
           document,
@@ -562,7 +580,7 @@ suite("Extension Test Suite", async () => {
       test("provide references for a suffix selector at a given position", async () => {
         const document = await workspace.openTextDocument(TestCssModulePath);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new ReferenceProvider();
         const result = await provider.provideReferences(
           document,
@@ -577,7 +595,7 @@ suite("Extension Test Suite", async () => {
       test("provide reference code lens for a selectors in a document", async () => {
         const document = await workspace.openTextDocument(TestCssModulePath);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new ReferenceCodeLensProvider();
         const result = await provider.provideCodeLenses(document, {
           isCancellationRequested: false,
@@ -587,12 +605,12 @@ suite("Extension Test Suite", async () => {
       test("provide references for a suffix selector in a document", async () => {
         const document = await workspace.openTextDocument(TestCssModulePath);
         await window.showTextDocument(document);
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const provider = new ReferenceCodeLensProvider();
         const lenses = await provider.provideCodeLenses(document, {
           isCancellationRequested: false,
         } as CancellationToken);
-        const result = provider.resolveCodeLens(
+        const result = await provider.resolveCodeLens(
           new ReferenceCodeLens(document, document.fileName, lenses[3].range)
         );
         assert.equal(result.command?.command, "editor.action.showReferences");
@@ -622,7 +640,7 @@ suite("TS Config path aliases", async () => {
         );
         const document = await workspace.openTextDocument(IndexComponent);
         await window.showTextDocument(document);
-        const diagnostics = await StorageInstance.bootStrap();
+        const diagnostics = await StorageInstance.experimental_BootStrap();
         assert.equal(diagnostics?.length, 0);
         StorageInstance.flushStorage();
       });
@@ -636,7 +654,7 @@ suite("TS Config path aliases", async () => {
         const document = await workspace.openTextDocument(IndexComponent);
         await window.showTextDocument(document);
 
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const definition = new DefnitionProvider();
         const position = new Position(6, 32);
         const result = await definition.provideDefinition(document, position);
@@ -654,7 +672,7 @@ suite("TS Config path aliases", async () => {
         const document = await workspace.openTextDocument(IndexComponent);
         await window.showTextDocument(document);
 
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const hover = new HoverProvider();
         const position = new Position(6, 33);
         const result = await hover.provideHover(document, position);
@@ -677,7 +695,7 @@ suite("TS Config path aliases", async () => {
       test("should not report an import diagnostics error on aliased module imports", async () => {
         const document = await workspace.openTextDocument(AppComponent);
         await window.showTextDocument(document);
-        const diagnostics = await StorageInstance.bootStrap();
+        const diagnostics = await StorageInstance.experimental_BootStrap();
         assert.equal(diagnostics?.length, 0);
         StorageInstance.flushStorage();
       });
@@ -686,7 +704,7 @@ suite("TS Config path aliases", async () => {
         const document = await workspace.openTextDocument(AppComponent);
         await window.showTextDocument(document);
 
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const definition = new DefnitionProvider();
         const position = new Position(17, 48);
         const result = await definition.provideDefinition(document, position);
@@ -699,7 +717,7 @@ suite("TS Config path aliases", async () => {
         const document = await workspace.openTextDocument(AppComponent);
         await window.showTextDocument(document);
 
-        await StorageInstance.bootStrap();
+        await StorageInstance.experimental_BootStrap();
         const hover = new HoverProvider();
         const position = new Position(17, 48);
         const result = await hover.provideHover(document, position);